src/main/java/com/havi/controller/BookController.java
package com.havi.controller;

import com.havi.service.BookService;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.*;

@Controller
public class BookController {

    @Autowired
    private BookService bookService;

    //@ResponseBody //view페이지가 아닌 응답값 그대로 반환하기 위해 사용
    @GetMapping("/books")
    public String getBookList(Model model) {
        model.addAttribute("bookList", bookService.getBookList());
        return "book";
    }
}